% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_ae_data}
\alias{assert_ae_data}
\title{Assertion of Adverse Event Data}
\usage{
assert_ae_data(data)
}
\arguments{
\item{data}{\code{data.frame} to be checked for \code{time_to_event} and \code{type_of_event} columns.}
}
\value{
None.
}
\description{
Custom assertion to check adverse event data sets.
}
\keyword{internal}
