% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Personal greeting}
\usage{
hello(name = "your name")
}
\arguments{
\item{name}{(\code{string})\cr a name, such as "john doe".}
}
\value{
(\code{string})\cr capitalized to title case.
}
\description{
Greet a person and appropriately capitalize their name.
}
\examples{
hello("james bond")
}
