% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savvyr-package.R
\docType{package}
\name{savvyr-package}
\alias{savvyr}
\alias{savvyr-package}
\title{savvyr: Survival Analysis for AdVerse Events with VarYing Follow-Up Times in R}
\description{
The SAVVY project is a consortium of academic and pharmaceutical industry partners that aims to improve the analyses of adverse event (AE) data in clinical trials through the use of survival techniques appropriately dealing with varying follow-up times and competing events. Although statistical methodologies have advanced, in AE analyses often the incidence proportion, the incidence density or a non-parametric Kaplan-Meier estimator are used, which either ignore censoring or competing events. This package contains functions to easily conduct the proposed improved AE analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/openpharma/savvyr/}
  \item Report bugs at \url{https://github.com/openpharma/savvyr/issues}
}

}
\author{
\strong{Maintainer}: Thomas Kuenzel \email{thomas.kuenzel@roche.com}

Authors:
\itemize{
  \item Kaspar Rufibach \email{kaspar.rufibach@roche.com}
  \item Regina Stegherr \email{regina.stegherr@uni-ulm.de}
  \item Daniel Sabanes Bove \email{daniel.sabanes_bove@roche.com}
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
  \item Universität Ulm [copyright holder, funder]
}

}
\keyword{internal}
