% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savvyr-package.R
\docType{package}
\name{savvyr-package}
\alias{savvyr}
\alias{savvyr-package}
\title{savvyr: R Package Template}
\description{
R package template with GitHub Actions workflows included.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/openpharma/savvyr/}
  \item Report bugs at \url{https://github.com/openpharma/savvyr/issues}
}

}
\author{
\strong{Maintainer}: openpharma \email{openpharma@example.com}

}
\keyword{internal}
