% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savvyr-package.R
\docType{package}
\name{savvyr-package}
\alias{savvyr}
\alias{savvyr-package}
\title{savvyr: Survival Analysis for AdVerse Events with VarYing Follow-Up Times in R}
\description{
TODO
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/openpharma/savvyr/}
  \item Report bugs at \url{https://github.com/openpharma/savvyr/issues}
}

}
\author{
\strong{Maintainer}: Thomas Kuenzel \email{thomas.kuenzel@roche.com}

Authors:
\itemize{
  \item Kaspar Rufibach \email{kaspar.rufibach@roche.com}
  \item Regina Stegherr \email{regina.stegherr@uni-ulm.de}
}

Other contributors:
\itemize{
  \item Daniel Sabanes Bove \email{daniel.sabanes_bove@roche.com} [contributor]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
  \item Universität Ulm [copyright holder, funder]
}

}
\keyword{internal}
